refactor: create text.comments-parser
