+++ text/comments-parser/README.md
@@ -1,0 +1,14 @@
+# @pnpm/text.comments-parser
+
+> Extracts and inserts comments from/to text
+
+## Installation
+
+```
+pnpm i @pnpm/text.comments-parser
+
+```
+
+## License
+
+[MIT](LICENSE)
