+++ text/comments-parser/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "@pnpm/text.comments-parser",
+  "description": "Extracts and inserts comments from/to text",
+  "version": "0.0.0",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib",
+    "!*.map"
+  ],
+  "keywords": [
+    "pnpm7"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">=14.6"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/main/text/comments-parser",
+  "scripts": {
+    "start": "tsc --watch",
+    "_test": "jest",
+    "test": "pnpm run compile && pnpm run _test",
+    "lint": "eslint src/**/*.ts test/**/*.ts",
+    "prepublishOnly": "pnpm run compile",
+    "compile": "tsc --build && pnpm run lint --fix"
+  },
+  "dependencies": {
+    "strip-comments-strings": "1.2.0"
+  },
+  "homepage": "https://github.com/pnpm/pnpm/blob/main/text/comments-parser#readme",
+  "funding": "https://opencollective.com/pnpm",
+  "devDependencies": {
+    "@pnpm/text.comments-parser": "workspace:*"
+  },
+  "exports": {
+    ".": "./lib/index.js"
+  }
+}
