+++ text/comments-parser/src/CommentSpecifier.ts
@@ -1,0 +1,9 @@
+export interface CommentSpecifier {
+  type: string
+  content: string
+  lineNumber: number
+  after?: string
+  on: string
+  whitespace: string
+  before?: string
+}
