+++ text/comments-parser/src/index.ts
@@ -1,0 +1,3 @@
+export * from './extractJson5Comments'
+export * from './insertJson5Comments'
+export * from './CommentSpecifier'
