+++ text/comments-parser/tsconfig.json
@@ -8,12 +8,5 @@
     "src/**/*.ts",
     "../../__typings__/**/*.d.ts"
   ],
-  "references": [
-    {
-      "path": "../../packages/types"
-    },
-    {
-      "path": "../../text/comments-parser"
-    }
-  ]
+  "references": []
 }
